create/update information refactoring

+++ client/src/components/updateCollectionElement/UpdateCollectionElement.jsx
@@ -3,6 +3,7 @@ import { useCollectionElementInformations } from '../contextProvider/CollectionE
 import { UpdateCollectionElementInformations } from '../updateCollectionElementInformations/UpdateCollectionElementInformations'
 import { useCollectionElementInformationsToUpdate } from '../contextProvider/CollectionElementInformationsToUpdateContextProvider'
 import { useCollectionElementInformationsToUpdateDelete } from '../contextProvider/CollectionElementInformationsToUpdateDeleteContextProvider'
+import { useCollectionElementInformationsToUpdateCreate } from '../contextProvider/CollectionElementInformationsToUpdateCreateContextProvider' 
 import '../collectionElement/CollectionElement.css'
 
 
@@ -10,8 +11,10 @@ export const UpdateCollectionElement = ({collectionElementDescription, collectio
 
 
   const {currentInformations}=useCollectionElementInformations()
-  const {currentInformationsToUpdate}=useCollectionElementInformationsToUpdate()
+  const {currentInformationsToUpdate, setCurrentInformationsToUpdate}=useCollectionElementInformationsToUpdate()
   const {currentInformationsToUpdateDelete}=useCollectionElementInformationsToUpdateDelete()
+  const {currentInformationsToUpdateCreate, setCurrentInformationsToUpdateCreate }=useCollectionElementInformationsToUpdateCreate()
+  const [currentInformationsToUpdateFinal, setCurrentInformationsToUpdateFinal]=useState([])
 
   const[collectionElementAttributesToUpdate, setCollectionElementAttributesToUpdate]=useState({
     collectionElementDescriptionToUpdate:collectionElementDescription,
@@ -25,19 +28,47 @@ export const UpdateCollectionElement = ({collectionElementDescription, collectio
 
 
   useEffect(()=>{
-    setCollectionElementAttributesToUpdate({
+    setCollectionElementAttributesToUpdate((prevCollectionElementAttributesToUpdate)=>({
+      ...prevCollectionElementAttributesToUpdate,
       collectionElementDescriptionToUpdate:collectionElementDescription,
       collectionElementEmailToUpdate: collectionElementEmail,
       collectionElementCookingToUpdate: collectionElementCooking,
       collectionElementRecommandationToUpdate: collectionElementRecommandation,
       collectionUID: collectionUID,
-      collectionElementInformationsToUpdate:currentInformationsToUpdate
-    })
-  },[collectionElementDescription, collectionElementEmail, collectionElementCooking, collectionElementRecommandation, collectionUID, currentInformationsToUpdate])
+      collectionElementInformationsToUpdate:currentInformationsToUpdateFinal
+    }))
+  },[collectionElementDescription, collectionElementEmail, collectionElementCooking, collectionElementRecommandation, collectionUID, currentInformationsToUpdateFinal])
 
 
-  const handleSubmit=async(e)=>{
+  const updateCollectionElement=async(e)=>{
     e.preventDefault()
+    console.log(currentInformations)
+    currentInformationsToUpdate.forEach((currentInformationToUpdate)=>{
+      const {collection_element_information_uid, collection_element_information_text}=currentInformationToUpdate
+      console.log(collection_element_information_uid)
+      console.log(collection_element_information_text)
+
+      if(!currentInformations.find((currentInformation)=>currentInformation.collection_element_information_uid==collection_element_information_uid)){
+        console.log('id not found')
+        setCurrentInformationsToUpdateCreate((prevInformationsToUpdateCreate)=>{
+          return [...prevInformationsToUpdateCreate,{
+            collection_element_information_uid:collection_element_information_uid,
+            collection_uid: currentInformations[0].collection_uid,
+            collection_element_information_text:collection_element_information_text
+          }]
+        })
+      } else {
+        console.log('else')
+        setCurrentInformationsToUpdateFinal((prevInformationsToUpdateFinal)=>{
+          return [...prevInformationsToUpdateFinal,{collection_element_information_uid: collection_element_information_uid, collection_element_information_text: collection_element_information_text}]
+        })
+      }
+    })
+
+    await new Promise((resolve) => {
+      setTimeout(resolve, 0);
+    });
+
     try {
       let response = await fetch('http://localhost:5000/admin/updateCollectionElementAttributes',{
         method: 'PUT',
@@ -65,54 +96,72 @@ export const UpdateCollectionElement = ({collectionElementDescription, collectio
           'Content-Type': 'application/json'
         },
         body:JSON.stringify({
-          informationsToUpdate:collectionElementAttributesToUpdate.collectionElementInformationsToUpdate
+          informationsToUpdate:currentInformationsToUpdateFinal
         })
       })
       let data=await response.json()
       console.log(data)
-      fetchCollectionElement()
-      fetchElementInformations(collectionUID)
     } catch (error) {
       console.error({message: error})  
     }
+    if(currentInformationsToUpdateDelete.length>0){
 
-    try {
-      console.log(currentInformationsToUpdateDelete)
-      let response= await fetch('http://localhost:5000/admin/deleteInformationInput',{
+      try {
+        console.log(currentInformationsToUpdateDelete)
+        let response= await fetch('http://localhost:5000/admin/deleteInformationInput',{
           method:'DELETE',
           headers: {
               'Content-Type': 'application/json'
           },
           body:JSON.stringify({
-              informationIdToDeleteList:currentInformationsToUpdateDelete
+            informationIdToDeleteList:currentInformationsToUpdateDelete
           })
-      })
-      let data=await response.json()
+        })
+        let data=await response.json()
         console.log(data)
-    } catch (error) {
+      } catch (error) {
         console.log(`information didn't get deleted => ${error}`)
         
+      }
     }
+    fetchCollectionElement()
+    fetchElementInformations(collectionUID)
+    // setCurrentInformationsToUpdateFinal([])
   }
+  
+  
+  
+
+  useEffect(()=>{
+    console.log('collection Element Attributes To Update =>',collectionElementAttributesToUpdate)
+  },[collectionElementAttributesToUpdate])
 
 
+  useEffect(()=>{
+    console.log('collection element information to Create ',currentInformationsToUpdateCreate)
+  },[currentInformationsToUpdateCreate])
 
 
-  // useEffect(()=>{
-  //   console.log(collectionElementAttributesToUpdate)
-  // },[collectionElementAttributesToUpdate])
+  useEffect(()=>{
+    console.log('collection element information to Update final ',currentInformationsToUpdateFinal)
+  },[currentInformationsToUpdateFinal])
 
-  const updateCollectionElement=(e)=>{
+    useEffect(()=>{
+    console.log('collection element information to Delete ',currentInformationsToUpdateDelete)
+  },[currentInformationsToUpdateDelete])
+
+  const handleChange=(e)=>{
     e.preventDefault()
-    setCollectionElementAttributesToUpdate({...collectionElementAttributesToUpdate,
+    setCollectionElementAttributesToUpdate((prevCollectionElementAttributesToUpdate)=>({
+      ...prevCollectionElementAttributesToUpdate,
       [e.target.name]:e.target.value
-    })
+    }))
   }
 
 
 
   return (
-    <form className='collectionElementInformationContainer' onSubmit={handleSubmit}>
+    <form className='collectionElementInformationContainer' onSubmit={updateCollectionElement}>
       <div className='collectionElementLeftContainer'>
         <label
           htmlFor='collectionElementDescription'>Collection Description  
@@ -120,7 +169,7 @@ export const UpdateCollectionElement = ({collectionElementDescription, collectio
         <input id='collectionElementDescription'
           type='text'
           value={collectionElementAttributesToUpdate.collectionElementDescriptionToUpdate}
-          onChange={updateCollectionElement}
+          onChange={handleChange}
           name='collectionElementDescriptionToUpdate'        
         />
       </div >
@@ -133,19 +182,19 @@ export const UpdateCollectionElement = ({collectionElementDescription, collectio
         <input id='collectionElementEmail'
           type='text'
           value={collectionElementAttributesToUpdate.collectionElementEmailToUpdate}
-          onChange={updateCollectionElement}
+          onChange={handleChange}
           name='collectionElementEmailToUpdate'    
         />
         <input id='collectionElementCooking'
           type='text'
           value={collectionElementAttributesToUpdate.collectionElementCookingToUpdate}
-          onChange={updateCollectionElement}
+          onChange={handleChange}
           name='collectionElementCookingToUpdate'        
         />
         <input id='collectionElementRecommandation'
           type='text'
           value={collectionElementAttributesToUpdate.collectionElementRecommandationToUpdate}
-          onChange={updateCollectionElement}
+          onChange={handleChange}
           name='collectionElementRecommandationToUpdate'        
         />
       </div>
